fix: Moving repeatable panel annotation to form-editor-support.js
